Separate out writing logic
